Audio Buttons

Buttons that can be used to interact with the howlerPlayer.

howlerPlayButton, howlerPauseButton, howlerPlayPauseButton and howlerStopButton will all be applied to the current track.

howlerBackButton and howlerForwardButton will change the track position by a specified amount of time.

howlerPreviousButton and howlerNextButton will play the previous/following track supplied to the player.

howlerVolumeDownButton and howlerVolumeUpButton will change the volume of the player by a specified percentage.

howlerButton is a customisable version of any of the above individual button.

howlerButton(id, button_type = HOWLER_BUTTON_TYPES, ...)

howlerPlayButton(id)

howlerPauseButton(id)

howlerPlayPauseButton(id)

howlerStopButton(id)

howlerBackButton(id, seek_change = 10)

howlerForwardButton(id, seek_change = 10)

howlerPreviousButton(id)

howlerNextButton(id)

howlerVolumeUpButton(id, volume_change = 0.1)

howlerVolumeDownButton(id, volume_change = 0.1)

howlerVolumeToggleButton(id)

Arguments

id
ID given to the button. For it to work with howlerPlayer, the ID must match that of the howlerPlayer.

button_type
Type of button to create. Available buttons are in the details, default set to play_pause.

...
Attributes/Inner tags added to the button

seek_change
Time (in seconds) to move forward/backward the track when clicked. Default is 10 seconds

volume_change
How much to change the volume by. Default is 10%.

Value

An HTML tag containing the audio button.

An additional input will be available in the server side in the form {id}_{button_type}. For example howlerBackButton("howler") will create an input element of input$howler_back. All of these will work in the same way as actionButton

Details

The following button_type are available to create:

play_pause

(default) Switch between playing and pausing the track

play

Resumes the current track

pause

Pauses the current track

stop

Stops current track, when played will start from beginning

previous,next

Switches to the previous/following track

volumedown,volumeup

Decreases/Increases the volume by 10% (If using howlerButton include the attribute `data-volume-change`)

back,forward

Seek forward/backwards 10s (If using howlerButton include the attribute `data-seek-change` with negative values to go backwards)

When using a play_pause button, the icon will toggle between the play and pause button depending on whether or not the track is playing.

Examples

if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    tile = "howler.js Player",
    useHowlerJS(),
    howlerPlayer("howler", "audio/sound.mp3"),
    howlerPreviousButton("howler"),
    howlerBackButton("howler"),
    howlerPlayPauseButton("howler"),
    howlerForwardButton("howler"),
    howlerNextButton("howler"),
    howlerVolumeDownButton("howler"),
    howlerVolumeUpButton("howler")
  )

  server <- function(input, output) {
  }

  shinyApp(ui, server)
}